requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
openpyxl==3.1.2
python-docx==1.1.2
openai>=1.3.0
google-generativeai>=0.3.0

